+++ src/hooks/screen/Home/Crypto/utils.ts
@@ -1,4 +1,4 @@
-import { _ChainAsset, _ChainInfo } from '@subwallet/chain-list/types';
+import { _ChainAsset, _ChainInfo } from '@soul-wallet/chain-list/types';
 import { AccountType } from 'types/ui-types';
 import {
   _getMultiChainAsset,