+++ src/messaging/index.ts
@@ -152,7 +152,7 @@ import {
   _ValidateCustomAssetRequest,
   _ValidateCustomAssetResponse,
 } from '@subwallet/extension-base/services/chain-service/types';
-import { _ChainAsset, _ChainInfo } from '@subwallet/chain-list/types';
+import { _ChainAsset, _ChainInfo } from '@soul-wallet/chain-list/types';
 import { AuthUrls } from '@subwallet/extension-base/services/request-service/types';
 import { _getKnownHashes } from 'utils/defaultChains';