+++ src/screens/Transaction/CancelUnstake/index.tsx
@@ -12,7 +12,7 @@ import useHandleSubmitTransaction from 'hooks/transaction/useHandleSubmitTransac
 import { ScrollView, View } from 'react-native';
 import { AccountSelectField } from 'components/Field/AccountSelect';
 import useGetAccountByAddress from 'hooks/screen/hooks/useGetAccountByAddress';
-import { _ChainInfo } from '@subwallet/chain-list/types';
+import { _ChainInfo } from '@soul-wallet/chain-list/types';
 import { AccountJson } from '@subwallet/extension-base/background/types';
 import { isSameAddress } from '@subwallet/extension-base/utils';
 import { accountFilterFunc } from 'screens/Transaction/helper/staking';