+++ src/screens/Transaction/SendFund/index.tsx
@@ -1,4 +1,4 @@
-import { _AssetRef, _ChainAsset, _ChainInfo, _MultiChainAsset } from '@subwallet/chain-list/types';
+import { _AssetRef, _ChainAsset, _ChainInfo, _MultiChainAsset } from '@soul-wallet/chain-list/types';
 import {
   _getAssetDecimals,
   _getOriginChainOfAsset,